third_party/WebKit/LayoutTests/platform/linux/http/tests/serviceworker/ServiceWorkerGlobalScope/registration-attribute-expected.txt

This is a testharness.js-based test.
FAIL Verify registration attributes on ServiceWorkerGlobalScopeassert_equals: Service Worker should respond to fetch expected "updatefound,install,statechange(installed),statechange(activating),activate,statechange(activated),fetch" but got "updatefound,install,activate,statechange(installed),statechange(activating),statechange(activated),fetch"
FAIL Verify registration attributes on ServiceWorkerGlobalScope of the newer workerpromise_test: Unhandled rejection with value: object "Error: wait_for_state must be passed a ServiceWorker"
Harness: the test ran to completion.
